Moves filter to a separate package
