+++ talaiot/src/main/kotlin/com/cdsap/talaiot/filter/Filter.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.configuration.filter
+package com.cdsap.talaiot.filter
 
 class StringFilter {
     var includes: Array<String>? = null
